Create rental preserve data methods
