+++ rentals.json
@@ -1,0 +1,12 @@
+[
+  {
+    "date": "2022/01/01",
+    "book": 0,
+    "person": 0
+  },
+  {
+    "date": "2022/01/01",
+    "book": 0,
+    "person": 0
+  }
+]
